+++ frontend/src/components/common/GroupArticleCard/styles.ts
