+++ frontend/src/components/common/StatCounter/index.tsx
