+++ frontend/src/pages/_app.tsx
@@ -2,7 +2,7 @@ import Head from 'next/head';
 import { AppProps } from 'next/app';
 import { QueryClientProvider, QueryClient } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
-import RouterTransition from '@components/RouterTransition';
+import RouterTransition from '@components/common/RouterTransition';
 import CommonStyles from '@styles/CommonStyles';
 import initMockApi from '@mocks/.';
 import { useEffect, useState } from 'react';
